pyproject.toml
[project]
name = "rss-alert"
version = "0.1.0"
description = "Sends an alert for new entries in an rss feed"
requires-python = ">=3.14"
dependencies = [
    "filelock>=3.25.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "tenacity>=9.1.4",
    "truststore>=0.10.4",
    "typer>=0.24.1",
    "xmltodict>=1.0.4",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
    "ty>=0.0.21",
]

[project.scripts]
rss-alert = "rss_alert.main:main"

[tool.uv]
package = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "S", "I"]
